+++ solver/src/sat_solver.rs
@@ -3,5 +3,5 @@ use crate::CNFClause;
 pub type Assignment = Vec<bool>;
 
 pub trait Solver {
-    fn solve(&self, clauses: impl Iterator<Item=CNFClause>, num_variables: u32) -> Option<Assignment>;
+    fn solve(&self, clauses: impl Iterator<Item=CNFClause>, num_variables: usize) -> Option<Assignment>;
 }
